Refactor search logic
